Histograma con porcentajes
Histograma con representacion en porcentajes completo

+++ Analizador_de_texto/Structures/CharList/CharNode.h
@@ -13,7 +13,7 @@ public:
     string character;
     CharNode *next;
     CharNode *last;
-    int counter;
+    float counter;
 public:
     explicit CharNode(string character);
 };
